Hotfix for the html5 video bug
This should be a temporary solution until we can find a better way to
fix this.
Since the `ParcelFileDescriptor` pipe always breaks when writing a video,
I've fixed the issue by just writing the video to the local file system
and then returning that local file to the ContentProvider.
This local file gets deleted when the App gets destroyed.