Move the JNI wrapper out of kiwixmobile and to kiwixlib package.
The JNI wrapping of kiwixlib should be move in the kiwixlib repository.
(And it is necessary because of kiwix.cpp who need to be compiled with
the lib)

The java part of the wrapper is also moved in the `org.kiwix.kiwixlib`
package.

The libkiwix.so and kiwixlib/*.java files are coming from the kiwix-lib
reposistory. They are copied here to simplify android development.
Those files should be replaced if any modification is made on the
kiwix-lib side. (and ideally, they should not be versioned here)