Add localized language and matching item count to language picker.
Shows the name of the language in that language for each choice in the
picker. Also shows number of matching items for that language.

A bit of refactoring, cleaning up the language population logic. Fixes
bug where after the initial language population & persisting no new
languages would be populated (if new item with that lang got downloaded).