I was irked by having a compile dependency on a testing library. Addressing the cause took several hours, it seemed to be there to address other side-effects. In the end I discovered we had incompatible library versions which weren't reconciled by Gradle unless the extraneous dependency was there.
The error was described in
https://sites.google.com/a/android.com/tools/tech-docs/new-build-system/user-guide#TOC-Resolving-conflicts-between-main-and-test-APK

The resolution was to align the version numbers for various Android
support libraries to a common value: 25.4.0 for now although there are
newer versions so we might want to migrate to the current releases soon.
I applied the concepts others use, which is to specify the version once
and use it throughout, see
https://segunfamisa.com/posts/android-gradle-extra-properties for a
concise example.

For info on the latest support and testing libraries see
https://developer.android.com/topic/libraries/support-library/revisions.html

https://developer.android.com/topic/libraries/testing-support-library/release-notes.html

We might also want to update our builds to the latest release of
Espresso.