Bug fix for #320 triggered when the library contained 1 item.
A sample library_zim.xml is available at
https://github.com/kiwix/testfiles/blob/master/libraryfiles/one_item_in_library/library_zim.xml

The bug seems to have been introduced inadvertently in
https://github.com/kiwix/kiwix-android/commit/b9930252e1a540366d7ecae5e0336a22497bc92c
which returned a null to `getView(...)`

I've left a key log statement as this was one clue (after checking
various other areas of the codebase) that indicated there might be a
problem related to the number of entries rather than the attributes of
the item. I aim to revise the format of the log message as part of a
wider update on logging.