Improved the UI: users can now understand the purpose of the WiFi option.
Duplicated the text that appears on the preferences in the summary so
that the user can actually read what it says. This particularly affects
smaller screened device. It might be worth revising the title, however
that'd then need any changes to be retranslated. This change isn't ideal
(duplicating content on the screen) however it seems practical and
helpful. You're welcome to help improve it.