Added additional logging. Temporarily added a default global crash handler for testing.
This global handler is not planned to be merged to master. It's to
enable testers to have a slightly better experience. Several of us are
working on a more complete crash handler that includes optional
reporting - that's more likely to be included if we decide and agree on
the benefits.

In terms of the additional logging, this is intended to help remote
testers send their android log file where we can find out more about
what's happening when they perform searches.