Added code to cope with the Android Permission prompts at runtime.
This enables the tests to run more reliably on devices rather than
relying on a person to grant the permission or for the DownloadTest to
have run (as it asked for permission in the code using the same
technique used here).

See https://github.com/kiwix/kiwix-android/issues/246 for more info.