Restyle the message in Notification Bar
The notification bar shows an option of "Play" instead of "Resume" upon starting
a download and pausing it. Fix the string resource and its corresponding reference.
Only the strings.xml file has to be edited as the data for other languages is auto generated
based on manual translations from translatewiki.net.