Target uploads to new dropbox address
SSH server which was used to receive file uploads (CI, nightly and release) has been
migrated to a new one on a different address.
Username, Key and paths are unchanged.
Most notable changes are the use of master.download.kiwix.org as the target in
replacement of mirror.download.kiwix.org (although it would still work) and the
Port to which SSH is listening on (30022 instead of 22)