Changed secret key as PLAYSTORE_JSON
We have updated the release.yml file to use the new PLAYSTORE_JSON secret key instead of the google_json secret key that was previously used to get credentials. As part of this update, we have renamed the google.json file to playstore.json for better naming consistency with the new secret key.